refactor cli storage helpers
